Headers for basic blocks in CFG dot graphs
This change adds separators for basic block names, which makes it
easier to find a basic block based on its name and separates it
from the code.

Currently there is also a chance that the basic block label will
be present twice, that is in case the basic block has explicit
numbering, this change fixes this bug.

Differential Revision: https://reviews.llvm.org/D154102