[clang-tidy] Container-size-empty fixed c++ version in tests to support string_literals operator
The goal of this PR is to properly implement the std::string_literals::operator""s that was added in commit (4001ae175cbe) but was later changed in commit (ba52a10fca6fc7b791894c584233db012def68a5).

The operator""s was added in c++14 but we were running tests under c++11 which would raise an error when compiling the file.

Reviewed By: PiotrZSL

Differential Revision: https://reviews.llvm.org/D158691