[AArch64] Fix schedmodel pre/post-index loads and stores for Neoverse V2
Fix the operand description: the update was in the wrong place. As a result the
latency of the update was modelled incorrectly, it wasn't available as early as
it should. This was visible in llvm-mca timeline views.

This fixes the problem for the Neoverse V2, but the problem also affects the
other Neoverse cores.

Patch by: Ricardo Jesus

Differential Revision: https://reviews.llvm.org/D159254