[llvm] Add CODE_OF_CONDUCT.md (#65816)
Add a CODE_OF_CONDUCT.md file to the root of the repository. The file
itself references the LLVM Community Code of Conduct. GitHub will
recognize this file and put a link to it to the right of the repository,
similar to the license and security policy, making the CoC easier to
discover.