[InlineAsm] break up Data Bitfield::Element into corresponding fields (#66297)
Use multiple Bitfield::Element's to provide more meaningful access to
the same slice of bits. "Data" used in different contexts is confusing.
Try to be as clear as possible, and in that vein, update the comment
block for this class, too.

This was suggested by @bwendling in
https://github.com/llvm/llvm-project/pull/65649#discussion_r1318962795
and @jyknight in

https://github.com/llvm/llvm-project/pull/65649#pullrequestreview-1620359094