[LLDB] Attempt to fix DumpDataExtractorTest
In 014c41d688f961cdb81cf92e6ba97100a49766a6 I tried to fix these tests,
but it seems that I needed to change TEST for TEST_F to make that work.
It's a pain that these failures don't repro on any of my machines, but I
verified thta the initialization code for the tests is invoked.