Reland "Add Documentation for Execution Results Handling in Clang-Repl (#65650)"
Original commit message: "

Add Documentation for Execution Results Handling in Clang-Repl (#65650)

    [clang-repl] Add Documentation for Execution Results Handling.

    This patch adds documentation for execution results handling in
    Clang-REPL with
    the below features:

    - Automatic Printf feature
    - Value Synthesis feature
    - Pretty Printing feature

    Continuing this work https://reviews.llvm.org/D156858 with this PR. I am
    issuing this patch on behalf of Saqib.

"

This reverts commit ac32d7b87f4d4b546eea96b9b722e88fdb3a5b49.

The new patch resolves the layering violation by simply not trying to look if
`dot` was installed. This is consistent with what we do other extensions.

After this patch we will need `sphinx.ext.graphviz` extension (part of the `dot`
package) which is available via anyway if we build the doxygen-based
documentation.