Run DCE after a LoopFlatten test to reduce spurious output [nfc]
LoopFlatten uses IndVars mechanics to widen induction variables, but then
fails to cleanup the dead instructions returned via the API.  This isn't
a functional problem, but means that this test is highly sensative to
indvars internals in an unhelpful way.  Use DCE to "normalize" the test
output to avoid false positives in a later change.