Update HPSCCFA.cpp to handle duplicated edges
In function addBufferNodeForFanIn(): changed fanInSuccs's type from vector to set because there are apparently edges that share the same predecessor and successor. In coremark benchmark, there is a case where the parent BB has a switch instruction where different switch cases jump to the same BB, introducing 2 edges from the extract same parrent BB and child BB. When adding buffer node in CFA, we need to remove the duplicates so we use *set* instead.

Signed-off-by: Yixue Zhao <felicitia2010@gmail.com>